optimized some code, made it prettier
